pyproject.toml
[project]
name = "testing-python-can"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-can>=4.6.1",
]
